Update enhancement_service tests
